Keep convertTreeToPushOptions and pull in the same file to avoid introducing more noise
